Ai/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
cors
motor
passlib[bcrypt]
bcrypt<4
python-jose[cryptography]
email-validator
